fix(site): navigation as tsx
