--- site/src/app/components/pass-sport-navigation/navigation.ts
@@ -1,14 +1,0 @@
-type NavigationItem = {
-  link: string;
-  text: string;
-};
-
-export const navigationItem: NavigationItem[] = [
-  {
-    link: '/v2/accueil',
-    text: 'Accueil',
-  },
-  { link: '/v2/tout-savoir-sur-le-pass-sport', text: 'Tout savoir sur le Pass Sport' },
-  { link: '/v2/une-question', text: 'Une question ?' },
-  { link: '/v2/trouver-un-club', text: 'Trouver un club' },
-];
